rename package reources to drawables
