+++ app/src/main/java/com/giraffe/tudeeapp/design_system/theme/Theme.kt
@@ -4,8 +4,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
 import com.giraffe.tudeeapp.design_system.color.LocalTudeeColors
 import com.giraffe.tudeeapp.design_system.color.TudeeColors
-import com.giraffe.tudeeapp.design_system.resources.LocalTudeeDrawables
-import com.giraffe.tudeeapp.design_system.resources.TudeeDrawables
+import com.giraffe.tudeeapp.design_system.drawables.LocalTudeeDrawables
+import com.giraffe.tudeeapp.design_system.drawables.TudeeDrawables
 import com.giraffe.tudeeapp.design_system.text_style.LocalTudeeTextStyle
 import com.giraffe.tudeeapp.design_system.text_style.TudeeTextStyle
